+++ recipes/module_pgsql.rb
@@ -1,6 +1,8 @@
 #
 # Author::  Joshua Timberman (<joshua@opscode.com>)
 # Author::  Seth Chisamore (<schisamo@opscode.com>)
+# Author::  Panagiotis Papadomitsos (<pj@ezgr.net>)
+#
 # Cookbook Name:: php
 # Recipe:: module_pgsql
 #
@@ -19,12 +21,9 @@
 # limitations under the License.
 #
 
-pkg = value_for_platform(
-  %w(centos redhat scientific fedora) => {
-    %w(5.0 5.1 5.2 5.3 5.4 5.5 5.6 5.7 5.8) => "php53-pgsql",
-    "default" => "php-pgsql"
-  },
-  "default" => "php5-pgsql"
+pkg = value_for_platform_family(
+    [ "rhel", "fedora" ] => { "default" => "php-pgsql" },
+    "debian" => { "default" => "php5-pgsql" }
 )
 
 package pkg do