+++ recipes/module_sqlite3.rb
@@ -1,10 +1,10 @@
 #
-# Author::  Joshua Timberman (<joshua@opscode.com>)
-# Author::  Seth Chisamore (<schisamo@opscode.com>)
-# Cookbook Name:: php
-# Recipe:: module_sqlite3
+# Author::  Panagiotis Papadomitsos (<pj@ezgr.net>)
 #
-# Copyright 2009-2011, Opscode, Inc.
+# Cookbook Name:: php
+# Recipe:: module_sqlite
+#
+# Copyright 2009-2012, Panagiotis Papadomitsos
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -19,12 +19,11 @@
 # limitations under the License.
 #
 
-case node['platform_family']
-when "rhel", "fedora"
-  #already there in centos, --with-pdo-sqlite=shared
-when "debian"
-  package "php5-sqlite" do
-    action :install
-  end
-end
+pkg = value_for_platform_family(
+    [ "rhel", "fedora" ] => { "default" => "php-sqlite" },
+    "debian" => { "default" => "php5-sqlite" }
+)
 
+package pkg do
+  action :install
+end