+++ recipes/source.rb
@@ -25,7 +25,7 @@ include_recipe "xml"
 include_recipe "mysql::client" if configure_options =~ /mysql/
 
 pkgs = value_for_platform(
-    ["centos","redhat","fedora", "scientific"] =>
+    ["centos","redhat","fedora"] =>
         {"default" => %w{ bzip2-devel libc-client-devel curl-devel freetype-devel gmp-devel libjpeg-devel krb5-devel libmcrypt-devel libpng-devel openssl-devel t1lib-devel mhash-devel }},
     [ "debian", "ubuntu" ] =>
         {"default" => %w{ libbz2-dev libc-client2007e-dev libcurl4-gnutls-dev libfreetype6-dev libgmp3-dev libjpeg62-dev libkrb5-dev libmcrypt-dev libpng12-dev libssl-dev libt1-dev }},
@@ -40,7 +40,7 @@ end
 
 version = node['php']['version']
 
-remote_file "#{Chef::Config[:file_cache_path]}/php-#{version}.tar.gz" do
+remote_file "#{Chef::Config['file_cache_path']}/php-#{version}.tar.gz" do
   source "#{node['php']['url']}/php-#{version}.tar.gz"
   checksum node['php']['checksum']
   mode "0644"
@@ -48,7 +48,7 @@ remote_file "#{Chef::Config[:file_cache_path]}/php-#{version}.tar.gz" do
 end
 
 bash "build php" do
-  cwd Chef::Config[:file_cache_path]
+  cwd Chef::Config['file_cache_path']
   code <<-EOF
   tar -zxvf php-#{version}.tar.gz
   (cd php-#{version} && ./configure #{configure_options})